+++ front/src/modules/ui/components/editable-cell/types/EditableCellDoubleText.tsx
@@ -1,4 +1,4 @@
-import { ReactElement, useEffect, useMemo, useState } from 'react';
+import { ReactElement, useMemo, useState } from 'react';
 
 import { InternalHotkeysScope } from '@/hotkeys/types/internal/InternalHotkeysScope';
 import { debounce } from '@/utils/debounce';
