+++ front/src/modules/ui/components/table/EntityTableHeader.tsx
@@ -1,5 +1,8 @@
 import { TableColumn } from '@/people/table/components/peopleColumns';
 
+import { ColumnHead } from './ColumnHead';
+import { SelectAllCheckbox } from './SelectAllCheckbox';
+
 export function EntityTableHeader({
   columns,
 }: {
@@ -8,6 +11,15 @@ export function EntityTableHeader({
   return (
     <thead>
       <tr>
+        <th
+          style={{
+            width: 30,
+            minWidth: 30,
+            maxWidth: 30,
+          }}
+        >
+          <SelectAllCheckbox />
+        </th>
         {columns.map((column) => (
           <th
             key={column.id.toString()}
@@ -17,7 +29,7 @@ export function EntityTableHeader({
               maxWidth: column.size,
             }}
           >
-            {column.title}
+            <ColumnHead viewName={column.title} viewIcon={column.icon} />
           </th>
         ))}
         <th></th>
