+++ front/src/pages/companies/CompanyTable.tsx
@@ -4,8 +4,9 @@ import { IconList } from '@tabler/icons-react';
 import {
   CompaniesSelectedSortType,
   defaultOrderBy,
-  useCompaniesQuery,
 } from '@/companies/services';
+import { companyColumns } from '@/companies/table/components/companyColumns';
+import { CompanyEntityTableData } from '@/companies/table/components/CompanyEntityTableData';
 import { reduceSortsToOrderBy } from '@/lib/filters-and-sorts/helpers';
 import { filtersScopedState } from '@/lib/filters-and-sorts/states/filtersScopedState';
 import { turnFilterIntoWhereClause } from '@/lib/filters-and-sorts/utils/turnFilterIntoWhereClause';
@@ -15,7 +16,6 @@ import { HooksEntityTable } from '@/ui/components/table/HooksEntityTable';
 import { TableContext } from '@/ui/tables/states/TableContext';
 import { CompanyOrderByWithRelationInput } from '~/generated/graphql';
 
-import { useCompaniesColumns } from './companies-columns';
 import { companiesFilters } from './companies-filters';
 import { availableSorts } from './companies-sorts';
 
@@ -30,27 +30,18 @@ export function CompanyTable() {
   const filters = useRecoilScopedValue(filtersScopedState, TableContext);
 
   const whereFilters = useMemo(() => {
-    if (!filters.length) return undefined;
-
     return { AND: filters.map(turnFilterIntoWhereClause) };
   }, [filters]) as any;
 
-  const companiesColumns = useCompaniesColumns();
-
-  const { data } = useCompaniesQuery(orderBy, whereFilters);
-
-  const companies = data?.companies ?? [];
-
   return (
     <>
+      <CompanyEntityTableData orderBy={orderBy} whereFilters={whereFilters} />
       <HooksEntityTable
-        numberOfColumns={companiesColumns.length}
-        numberOfRows={companies.length}
-        availableTableFilters={companiesFilters}
+        numberOfColumns={companyColumns.length}
+        availableFilters={companiesFilters}
       />
       <EntityTable
-        data={companies}
-        columns={companiesColumns}
+        columns={companyColumns}
         viewName="All Companies"
         viewIcon={<IconList size={16} />}
         availableSorts={availableSorts}
