+++ front/src/pages/people/PeopleTable.tsx
@@ -5,14 +5,15 @@ import { defaultOrderBy } from '@/companies/services';
 import { reduceSortsToOrderBy } from '@/lib/filters-and-sorts/helpers';
 import { filtersScopedState } from '@/lib/filters-and-sorts/states/filtersScopedState';
 import { turnFilterIntoWhereClause } from '@/lib/filters-and-sorts/utils/turnFilterIntoWhereClause';
-import { PeopleSelectedSortType, usePeopleQuery } from '@/people/services';
+import { PeopleEntityTableData } from '@/people/components/PeopleEntityTableData';
+import { PeopleSelectedSortType } from '@/people/services';
+import { peopleColumns } from '@/people/table/components/peopleColumns';
 import { useRecoilScopedValue } from '@/recoil-scope/hooks/useRecoilScopedValue';
 import { EntityTable } from '@/ui/components/table/EntityTable';
 import { HooksEntityTable } from '@/ui/components/table/HooksEntityTable';
 import { TableContext } from '@/ui/tables/states/TableContext';
 import { PersonOrderByWithRelationInput } from '~/generated/graphql';
 
-import { usePeopleColumns } from './people-columns';
 import { peopleFilters } from './people-filters';
 import { availableSorts } from './people-sorts';
 
@@ -30,21 +31,14 @@ export function PeopleTable() {
     return { AND: filters.map(turnFilterIntoWhereClause) };
   }, [filters]) as any;
 
-  const peopleColumns = usePeopleColumns();
-
-  const { data } = usePeopleQuery(orderBy, whereFilters);
-
-  const people = data?.people ?? [];
-
   return (
     <>
+      <PeopleEntityTableData orderBy={orderBy} whereFilters={whereFilters} />
       <HooksEntityTable
         numberOfColumns={peopleColumns.length}
-        numberOfRows={people.length}
-        availableTableFilters={peopleFilters}
+        availableFilters={peopleFilters}
       />
       <EntityTable
-        data={people}
         columns={peopleColumns}
         viewName="All People"
         viewIcon={<IconList size={16} />}
